+++ packages/ui/src/index.ts
@@ -3,6 +3,7 @@ import "./assets/css/index.css";
 import { Accordion } from "./Accordion";
 import { SubmitButton } from "./Buttons";
 import { SubPane } from "./components/SubPane";
+import Divider from "./Divider";
 import DropdownMenu, { DropdownMenuProperties } from "./DropdownMenu";
 import LoadingIcon from "./LoadingIcon";
 import LoadingPage from "./LoadingPage";
@@ -11,9 +12,11 @@ import ResponsiveMenu from "./ResponsiveMenu";
 import { GoogleButton } from "./SSOButtons";
 import { TabbedPanel } from "./TabbedPanel";
 export * from "./Table";
+export * from "./Utils";
 
 export {
   Accordion,
+  Divider,
   DropdownMenu,
   GoogleButton,
   LoadingIcon,