+++ packages/user/src/layouts/UserEnabledSidebarLayout.tsx
@@ -2,7 +2,10 @@ import {
   AppHeader,
   CollapsibleSidebarLayout,
   Logo,
+  MainMenu,
+  Sidebar,
 } from "@dzangolab/react-layout";
+import { ResponsiveMenu, useMediaQuery } from "@dzangolab/react-ui";
 
 import DropdownUserMenu from "../components/DropdownUserMenu";
 import UserMenu from "../components/UserMenu";
@@ -13,8 +16,12 @@ import { UserMenuItemType } from "../types";
 interface Properties {
   children: React.ReactNode;
   footer?: React.ReactNode;
+  mainMenuRoutes: {
+    name: string;
+    route: string;
+  }[];
   onLogout?: () => void;
-  sidebar?: React.ReactNode;
+  customSidebar?: React.ReactNode;
   userMenu?: UserMenuItemType[];
   userMenuCollapsedIcon?: React.ReactNode;
   userMenuExpandIcon?: React.ReactNode;
@@ -24,14 +31,20 @@ interface Properties {
 const UserEnabledSidebarLayout: React.FC<Properties> = (properties) => {
   const { layout: layoutConfig, user: userConfig } = useConfig();
   const { user } = useUser();
+  const isSmallScreen = useMediaQuery("(max-width: 576px)");
 
   const home = getHomeRoute(user, layoutConfig, userConfig);
 
   const {
     children,
     footer,
+    mainMenuRoutes,
     onLogout,
-    sidebar,
+    customSidebar = (
+      <Sidebar>
+        <ResponsiveMenu routes={mainMenuRoutes} horizontal />
+      </Sidebar>
+    ),
     userMenu,
     userMenuCollapsedIcon,
     userMenuExpandIcon,
@@ -44,7 +57,7 @@ const UserEnabledSidebarLayout: React.FC<Properties> = (properties) => {
       footer={footer}
       header={
         <AppHeader
-          mainMenu={null}
+          mainMenu={isSmallScreen ? <MainMenu routes={mainMenuRoutes} /> : null}
           userMenu={
             <UserMenu
               authenticatedUserMenu={
@@ -61,7 +74,7 @@ const UserEnabledSidebarLayout: React.FC<Properties> = (properties) => {
           logo={<Logo source={layoutConfig?.logo} route={home} />}
         />
       }
-      sidebar={user ? sidebar : null}
+      sidebar={user && !isSmallScreen ? customSidebar : null}
     />
   );
 };